+++ lib/class/channel.class.php
@@ -56,15 +56,12 @@ class Channel extends database_object implements media, library_item
 
     /**
      * Constructor
+     * @param integer $channel_id
      */
-    public function __construct($id = 0)
+    public function __construct($channel_id)
     {
-        if (!$id) {
-            return true;
-        }
-
         /* Get the information from the db */
-        $info = $this->get_info($id);
+        $info = $this->get_info($channel_id);
 
         // Foreach what we've got
         foreach ($info as $key => $value) {