+++ lib/class/share.class.php
@@ -48,14 +48,10 @@ class Share extends database_object
     /**
      * Constructor
      */
-    public function __construct($id = 0)
+    public function __construct($share_id)
     {
-        if (!$id) {
-            return true;
-        }
-
         /* Get the information from the db */
-        $info = $this->get_info($id);
+        $info = $this->get_info($share_id);
 
         // Foreach what we've got
         foreach ($info as $key => $value) {