+++ lib/class/subsonic_api.class.php
@@ -40,17 +40,6 @@ class Subsonic_Api
     {
     }
 
-    public static function check_version($input, $version = "1.0.0", $addheader = false)
-    {
-        // We cannot check client version unfortunately. Most Subsonic client sent a dummy client version...
-        /* if (version_compare($input['v'], $version) < 0) {
-          ob_end_clean();
-          if ($addheader) self::setHeader($input['f']);
-          self::apiOutput($input, Subsonic_XML_Data::createError(Subsonic_XML_Data::SSERROR_APIVERSION_CLIENT));
-          return false;
-          } */
-    }
-
     /**
      * @param string $parameter
      */
@@ -346,8 +335,6 @@ class Subsonic_Api
      */
     public static function getlicense($input)
     {
-        self::check_version($input);
-
         $response = Subsonic_XML_Data::createSuccessResponse();
         Subsonic_XML_Data::addLicense($response);
         self::apiOutput($input, $response);
@@ -360,8 +347,6 @@ class Subsonic_Api
      */
     public static function getmusicfolders($input)
     {
-        self::check_version($input);
-
         $r = Subsonic_XML_Data::createSuccessResponse();
         Subsonic_XML_Data::addMusicFolders($r, Catalog::get_catalogs());
         self::apiOutput($input, $r);
@@ -374,7 +359,6 @@ class Subsonic_Api
      */
     public static function getindexes($input)
     {
-        self::check_version($input);
         set_time_limit(300);
 
         $musicFolderId   = $input['musicFolderId'];
@@ -430,8 +414,6 @@ class Subsonic_Api
      */
     public static function getmusicdirectory($input)
     {
-        self::check_version($input);
-
         $id = self::check_parameter($input, 'id');
 
         $r = Subsonic_XML_Data::createSuccessResponse();
@@ -454,8 +436,6 @@ class Subsonic_Api
      */
     public static function getgenres($input)
     {
-        self::check_version($input, "1.9.0");
-
         $r = Subsonic_XML_Data::createSuccessResponse();
         Subsonic_XML_Data::addGenres($r, Tag::get_tags('song'));
         self::apiOutput($input, $r);
@@ -468,8 +448,6 @@ class Subsonic_Api
      */
     public static function getartists($input)
     {
-        self::check_version($input, "1.7.0");
-
         $r       = Subsonic_XML_Data::createSuccessResponse();
         $artists = Catalog::get_artists(Catalog::get_catalogs());
         Subsonic_XML_Data::addArtistsRoot($r, $artists, true);
@@ -483,8 +461,6 @@ class Subsonic_Api
      */
     public static function getartist($input)
     {
-        self::check_version($input, "1.7.0");
-
         $artistid = self::check_parameter($input, 'id');
 
         $artist = new Artist(Subsonic_XML_Data::getAmpacheId($artistid));
@@ -504,8 +480,6 @@ class Subsonic_Api
      */
     public static function getalbum($input)
     {
-        self::check_version($input, "1.7.0");
-
         $albumid = self::check_parameter($input, 'id');
 
         $addAmpacheInfo = ($input['ampache'] == "1");
@@ -528,8 +502,6 @@ class Subsonic_Api
      */
     public static function getvideos($input)
     {
-        self::check_version($input, "1.7.0");
-
         $response = Subsonic_XML_Data::createSuccessResponse();
         $videos   = Catalog::get_videos();
         Subsonic_XML_Data::addVideos($response, $videos);
@@ -543,8 +515,6 @@ class Subsonic_Api
      */
     public static function getalbumlist($input, $elementName = "albumList")
     {
-        self::check_version($input, "1.2.0");
-
         $type = self::check_parameter($input, 'type');
 
         $size          = $input['size'];
@@ -559,86 +529,62 @@ class Subsonic_Api
             $catalogs[] = $musicFolderId;
         }
 
-        $r            = Subsonic_XML_Data::createSuccessResponse();
+        $response     = Subsonic_XML_Data::createSuccessResponse();
         $errorOccured = false;
         $albums       = array();
 
-        if ($type == "random") {
-            $albums = Album::get_random($size);
-        } else {
-            if ($type == "newest") {
+        switch ($type) {
+            case "random":
+                $albums = Album::get_random($size);
+                break;
+            case "newest":
                 $albums = Stats::get_newest("album", $size, $offset, $musicFolderId);
-            } else {
-                if ($type == "highest") {
-                    $albums = Rating::get_highest("album", $size, $offset);
-                } else {
-                    if ($type == "frequent") {
-                        $albums = Stats::get_top("album", $size, '', $offset);
-                    } else {
-                        if ($type == "recent") {
-                            $albums = Stats::get_recent("album", $size, $offset);
-                        } else {
-                            if ($type == "starred") {
-                                $albums = Userflag::get_latest('album', null, $size);
-                            } else {
-                                if ($type == "alphabeticalByName") {
-                                    $albums = Catalog::get_albums($size, $offset, $catalogs);
-                                } else {
-                                    if ($type == "alphabeticalByArtist") {
-                                        $albums = Catalog::get_albums_by_artist($size, $offset, $catalogs);
-                                    } else {
-                                        if ($type == "byYear") {
-                                            $fromYear = $input['fromYear'];
-                                            $toYear   = $input['toYear'];
+                break;
+            case "highest":
+                $albums = Rating::get_highest("album", $size, $offset);
+                break;
+            case "frequent":
+                $albums = Stats::get_top("album", $size, '', $offset);
+                break;
+            case "recent":
+                $albums = Stats::get_recent("album", $size, $offset);
+                break;
+            case "starred":
+                $albums = Userflag::get_latest('album', null, $size);
+                break;
+            case "alphabeticalByName":
+                $albums = Catalog::get_albums($size, $offset, $catalogs);
+                break;
+            case "alphabeticalByArtist":
+                $albums = Catalog::get_albums_by_artist($size, $offset, $catalogs);
+                break;
+            case "byYear":
+                $fromYear = $input['fromYear'];
+                $toYear   = $input['toYear'];
 
-                                            if ($fromYear || $toYear) {
-                                                $search           = array();
-                                                $search['limit']  = $size;
-                                                $search['offset'] = $offset;
-                                                $search['type']   = "album";
-                                                $count            = 0;
-                                                if ($fromYear) {
-                                                    $search['rule_' . $count . '_input']    = $fromYear;
-                                                    $search['rule_' . $count . '_operator'] = 0;
-                                                    $search['rule_' . $count . '']          = "year";
-                                                    ++$count;
-                                                }
-                                                if ($toYear) {
-                                                    $search['rule_' . $count . '_input']    = $toYear;
-                                                    $search['rule_' . $count . '_operator'] = 1;
-                                                    $search['rule_' . $count . '']          = "year";
-                                                    ++$count;
-                                                }
-
-                                                $query  = new Search(null, 'album');
-                                                $albums = $query->run($search);
-                                            }
-                                        } else {
-                                            if ($type == "byGenre") {
-                                                $genre = self::check_parameter($input, 'genre');
-
-                                                $tag_id = Tag::tag_exists($genre);
-                                                if ($tag_id) {
-                                                    $albums = Tag::get_tag_objects('album', $tag_id, $size, $offset);
-                                                }
-                                            } else {
-                                                $r            = Subsonic_XML_Data::createError(Subsonic_XML_Data::SSERROR_GENERIC, "Invalid list type: " . scrub_out($type));
-                                                $errorOccured = true;
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
+                if ($fromYear || $toYear) {
+                    $search = Search::year_search($fromYear, $toYear, $size, $offset);
+                    $query  = new Search(null, 'album');
+                    $albums = $query->run($search);
                 }
-            }
+                break;
+            case "byGenre":
+                $genre = self::check_parameter($input, 'genre');
+
+                $tag_id = Tag::tag_exists($genre);
+                if ($tag_id) {
+                    $albums = Tag::get_tag_objects('album', $tag_id, $size, $offset);
+                }
+                break;
+            default:
+                $response     = Subsonic_XML_Data::createError(Subsonic_XML_Data::SSERROR_GENERIC, "Invalid list type: " . scrub_out($type));
+                $errorOccured = true;
         }
 
         if (!$errorOccured) {
-            Subsonic_XML_Data::addAlbumList($r, $albums, $elementName);
+            Subsonic_XML_Data::addAlbumList($response, $albums, $elementName);
         }
-        self::apiOutput($input, $r);
+        self::apiOutput($input, $response);
     }
 
     /**
@@ -647,7 +593,6 @@ class Subsonic_Api
      */
     public static function getalbumlist2($input)
     {
-        self::check_version($input, "1.7.0");
         self::getAlbumList($input, "albumList2");
     }
 
@@ -658,8 +603,6 @@ class Subsonic_Api
      */
     public static function getrandomsongs($input)
     {
-        self::check_version($input, "1.2.0");
-
         $size = $input['size'];
         if (!$size) {
             $size = 10;
@@ -716,7 +659,7 @@ class Subsonic_Api
             ++$count;
         }
         if ($count > 0) {
-            $songs = Random::advanced("song", $search);
+            $songs = Random::advanced('song', $search);
         } else {
             $songs = Random::get_default($size);
         }
@@ -733,13 +676,34 @@ class Subsonic_Api
      */
     public static function getsong($input)
     {
-        self::check_version($input, "1.7.0");
-
         $songid = self::check_parameter($input, 'id');
         $r      = Subsonic_XML_Data::createSuccessResponse();
         $song   = Subsonic_XML_Data::getAmpacheId($songid);
         Subsonic_XML_Data::addSong($r, $song);
         self::apiOutput($input, $r, array('musicFolder', 'artist', 'child', 'playlist', 'album'));
+    }
+
+    /**
+     * getTopSongs
+     * Get most popular songs for a given artist.
+     * Takes the genre with optional count and offset in parameters.
+     */
+    public static function gettopsongs($input)
+    {
+        $artist_id = self::check_parameter($input, 'artist');
+        $artist    = new Artist(Subsonic_XML_Data::getAmpacheId($artist_id));
+        $count     = (int) $input['count'];
+        if ($count <= 0) {
+            $count = 50;
+        }
+        if ($artist->id) {
+            $songs = Artist::get_top_songs($artist->id, $count);
+        } else {
+            $songs = array();
+        }
+        $response = Subsonic_XML_Data::createSuccessResponse();
+        Subsonic_XML_Data::addTopSongs($response, $songs);
+        self::apiOutput($input, $response);
     }
 
     /**
@@ -749,8 +713,6 @@ class Subsonic_Api
      */
     public static function getsongsbygenre($input)
     {
-        self::check_version($input, "1.9.0");
-
         $genre  = self::check_parameter($input, 'genre');
         $count  = $input['count'];
         $offset = $input['offset'];
@@ -773,8 +735,6 @@ class Subsonic_Api
      */
     public static function getnowplaying($input)
     {
-        self::check_version($input);
-
         $data = Stream::get_now_playing();
         $r    = Subsonic_XML_Data::createSuccessResponse();
         Subsonic_XML_Data::addNowPlaying($r, $data);
@@ -788,8 +748,6 @@ class Subsonic_Api
      */
     public static function search2($input, $elementName = "searchResult2")
     {
-        self::check_version($input, "1.2.0");
-
         $query    = self::check_parameter($input, 'query');
         $artists  = array();
         $albums   = array();
@@ -860,7 +818,6 @@ class Subsonic_Api
      */
     public static function search3($input)
     {
-        self::check_version($input, "1.7.0");
         self::search2($input, "searchResult3");
     }
 
@@ -871,8 +828,6 @@ class Subsonic_Api
      */
     public static function getplaylists($input)
     {
-        self::check_version($input);
-
         $r        = Subsonic_XML_Data::createSuccessResponse();
         $username = $input['username'];
 
@@ -897,8 +852,6 @@ class Subsonic_Api
      */
     public static function getplaylist($input)
     {
-        self::check_version($input);
-
         $playlistid = self::check_parameter($input, 'id');
 
         $r = Subsonic_XML_Data::createSuccessResponse();
@@ -919,8 +872,6 @@ class Subsonic_Api
      */
     public static function createplaylist($input)
     {
-        self::check_version($input, "1.2.0");
-
         $playlistId = $input['playlistId'];
         $name       = $input['name'];
         $songId     = $input['songId'];
@@ -985,8 +936,6 @@ class Subsonic_Api
      */
     public static function updateplaylist($input)
     {
-        self::check_version($input, "1.7.0");
-
         $playlistId = self::check_parameter($input, 'playlistId');
 
         $name    = $input['name'];
@@ -1013,8 +962,6 @@ class Subsonic_Api
      */
     public static function deleteplaylist($input)
     {
-        self::check_version($input, "1.2.0");
-
         $playlistId = self::check_parameter($input, 'id');
 
         if (Subsonic_XML_Data::isSmartPlaylist($playlistId)) {
@@ -1036,8 +983,6 @@ class Subsonic_Api
      */
     public static function stream($input)
     {
-        self::check_version($input, "1.0.0", true);
-
         $fileid = self::check_parameter($input, 'id', true);
 
         $maxBitRate            = $input['maxBitRate'];
@@ -1078,8 +1023,6 @@ class Subsonic_Api
      */
     public static function download($input)
     {
-        self::check_version($input, "1.0.0", true);
-
         $fileid = self::check_parameter($input, 'id', true);
 
         $url = Song::play_url(Subsonic_XML_Data::getAmpacheId($fileid), '&action=download' . '&client=' . rawurlencode($input['c']) . '&noscrobble=1', 'api', function_exists('curl_version'));
@@ -1093,8 +1036,6 @@ class Subsonic_Api
      */
     public static function hls($input)
     {
-        self::check_version($input, "1.7.0", true);
-
         $fileid = self::check_parameter($input, 'id', true);
 
         $bitRate = $input['bitRate'];
@@ -1124,8 +1065,6 @@ class Subsonic_Api
      */
     public static function getcoverart($input)
     {
-        self::check_version($input, "1.0.0", true);
-
         $id   = self::check_parameter($input, 'id', true);
         $size = $input['size'];
 
@@ -1176,8 +1115,6 @@ class Subsonic_Api
      */
     public static function setrating($input)
     {
-        self::check_version($input, "1.6.0");
-
         $object_id = self::check_parameter($input, 'id');
         $rating    = $input['rating'];
 
@@ -1213,8 +1150,6 @@ class Subsonic_Api
      */
     public static function getstarred($input, $elementName = "starred")
     {
-        self::check_version($input, "1.7.0");
-
         $user_id = Core::get_global('user')->id;
 
         $response = Subsonic_XML_Data::createSuccessResponse();
@@ -1239,8 +1174,6 @@ class Subsonic_Api
      */
     public static function star($input)
     {
-        self::check_version($input, "1.7.0");
-
         self::_setStar($input, true);
     }
 
@@ -1252,8 +1185,6 @@ class Subsonic_Api
      */
     public static function unstar($input)
     {
-        self::check_version($input, "1.7.0");
-
         self::_setStar($input, false);
     }
 
@@ -1330,8 +1261,6 @@ class Subsonic_Api
      */
     public static function getuser($input)
     {
-        self::check_version($input, "1.3.0");
-
         $username = self::check_parameter($input, 'username');
 
         if (Core::get_global('user')->access >= 100 || Core::get_global('user')->username == $username) {
@@ -1356,8 +1285,6 @@ class Subsonic_Api
      */
     public static function getusers($input)
     {
-        self::check_version($input, "1.7.0");
-
         if (Core::get_global('user')->access >= 100) {
             $response     = Subsonic_XML_Data::createSuccessResponse();
             $users        = User::get_valid_users();
@@ -1410,8 +1337,6 @@ class Subsonic_Api
      */
     public static function getinternetradiostations($input)
     {
-        self::check_version($input, "1.9.0");
-
         $response      = Subsonic_XML_Data::createSuccessResponse();
         $radios        = Live_Stream::get_all_radios();
         Subsonic_XML_Data::addRadios($response, $radios);
@@ -1425,8 +1350,6 @@ class Subsonic_Api
      */
     public static function getshares($input)
     {
-        self::check_version($input, "1.6.0");
-
         $response      = Subsonic_XML_Data::createSuccessResponse();
         $shares        = Share::get_share_list();
         Subsonic_XML_Data::addShares($response, $shares);
@@ -1440,8 +1363,6 @@ class Subsonic_Api
      */
     public static function createshare($input)
     {
-        self::check_version($input, "1.6.0");
-
         $id          = self::check_parameter($input, 'id');
         $description = $input['description'];
 
@@ -1487,8 +1408,6 @@ class Subsonic_Api
      */
     public static function deleteshare($input)
     {
-        self::check_version($input, "1.6.0");
-
         $id = self::check_parameter($input, 'id');
 
         if (AmpConfig::get('share')) {
@@ -1511,8 +1430,6 @@ class Subsonic_Api
      */
     public static function updateshare($input)
     {
-        self::check_version($input, "1.6.0");
-
         $id          = self::check_parameter($input, 'id');
         $description = $input['description'];
 
@@ -1561,8 +1478,6 @@ class Subsonic_Api
      */
     public static function createuser($input)
     {
-        self::check_version($input, "1.1.0");
-
         $username     = self::check_parameter($input, 'username');
         $password     = self::check_parameter($input, 'password');
         $email        = self::check_parameter($input, 'email');
@@ -1610,14 +1525,76 @@ class Subsonic_Api
     }
 
     /**
+     * updateUser
+     * Update an existing user.
+     * Takes the username with optional parameters.
+     */
+    public static function updateuser($input)
+    {
+        $username = self::check_parameter($input, 'username');
+        $password = $input['password'];
+        $email    = $input['email'];
+        //$ldapAuthenticated = $input['ldapAuthenticated'];
+        $adminRole    = ($input['adminRole'] == 'true');
+        $downloadRole = ($input['downloadRole'] == 'true');
+        $uploadRole   = ($input['uploadRole'] == 'true');
+        $coverArtRole = ($input['coverArtRole'] == 'true');
+        $shareRole    = ($input['shareRole'] == 'true');
+        //$musicfolderid = $input['musicFolderId'];
+        $maxbitrate = $input['maxBitRate'];
+
+        if (Access::check('interface', 100)) {
+            $access = 25;
+            if ($adminRole) {
+                $access = 100;
+            } elseif ($coverArtRole) {
+                $access = 75;
+            }
+            // identify the user to modify
+            $user_id  = User::get_from_username($username);
+            $user     = new User($user_id);
+
+            if ($user_id > 0) {
+                // update password
+                if ($password) {
+                    $password = self::decrypt_password($password);
+                    $user->update_password($password);
+                }
+                // update e-mail
+                if (Mailer::validate_address($email)) {
+                    $user->update_email($email);
+                }
+                // set preferences
+                if ($downloadRole) {
+                    Preference::update('download', $user_id, 1);
+                }
+                if ($uploadRole) {
+                    Preference::update('allow_upload', $user_id, 1);
+                }
+                if ($shareRole) {
+                    Preference::update('share', $user_id, 1);
+                }
+                if ((int) $maxbitrate > 0) {
+                    Preference::update('transcode_bitrate', $user_id, $maxbitrate);
+                }
+                $response = Subsonic_XML_Data::createSuccessResponse();
+            } else {
+                $response = Subsonic_XML_Data::createError(Subsonic_XML_Data::SSERROR_DATA_NOTFOUND);
+            }
+        } else {
+            $response = Subsonic_XML_Data::createError(Subsonic_XML_Data::SSERROR_UNAUTHORIZED);
+        }
+
+        self::apiOutput($input, $response);
+    }
+
+    /**
      * deleteUser
      * Delete an existing user.
      * Takes the username in parameter.
      */
     public static function deleteuser($input)
     {
-        self::check_version($input, "1.3.0");
-
         $username = self::check_parameter($input, 'username');
         if (Access::check('interface', 100)) {
             $user = User::get_from_username($username);
@@ -1641,8 +1618,6 @@ class Subsonic_Api
      */
     public static function changepassword($input)
     {
-        self::check_version($input, "1.1.0");
-
         $username     = self::check_parameter($input, 'username');
         $inp_pass     = self::check_parameter($input, 'password');
         $password     = self::decrypt_password($inp_pass);
@@ -1669,7 +1644,6 @@ class Subsonic_Api
      */
     public static function jukeboxcontrol($input)
     {
-        self::check_version($input, "1.2.0");
         $action = self::check_parameter($input, 'action');
         $id     = $input['id'];
         $gain   = $input['gain'];
@@ -1766,8 +1740,6 @@ class Subsonic_Api
      */
     public static function scrobble($input)
     {
-        self::check_version($input, "1.5.0");
-
         $id         = self::check_parameter($input, 'id');
         $submission = $input['submission'];
         //$time = $input['time'];
@@ -1803,8 +1775,6 @@ class Subsonic_Api
      */
     public static function getlyrics($input)
     {
-        self::check_version($input, "1.2.0");
-
         $artist = $input['artist'];
         $title  = $input['title'];
 
@@ -1944,7 +1914,6 @@ class Subsonic_Api
      */
     public static function getpodcasts($input)
     {
-        self::check_version($input, "1.6.0");
         $id              = $input['id'];
         $includeEpisodes = isset($input['includeEpisodes']) ? $input['includeEpisodes'] : true;
 
@@ -1975,7 +1944,6 @@ class Subsonic_Api
      */
     public static function getnewestpodcasts($input)
     {
-        self::check_version($input, "1.13.0");
         //$count = $input['count'] ?: 20; // Seems to be useless code
 
         if (AmpConfig::get('podcast')) {
@@ -1995,8 +1963,6 @@ class Subsonic_Api
      */
     public static function refreshpodcasts($input)
     {
-        self::check_version($input, "1.9.0");
-
         if (AmpConfig::get('podcast') && Access::check('interface', 75)) {
             $podcasts = Catalog::get_podcasts();
             foreach ($podcasts as $podcast) {
@@ -2016,7 +1982,6 @@ class Subsonic_Api
      */
     public static function createpodcastchannel($input)
     {
-        self::check_version($input, "1.9.0");
         $url = self::check_parameter($input, 'url');
 
         if (AmpConfig::get('podcast') && Access::check('interface', 75)) {
@@ -2046,7 +2011,6 @@ class Subsonic_Api
      */
     public static function deletepodcastchannel($input)
     {
-        self::check_version($input, "1.9.0");
         $id = self::check_parameter($input, 'id');
 
         if (AmpConfig::get('podcast') && Access::check('interface', 75)) {
@@ -2073,7 +2037,6 @@ class Subsonic_Api
      */
     public static function deletepodcastepisode($input)
     {
-        self::check_version($input, "1.9.0");
         $id = self::check_parameter($input, 'id');
 
         if (AmpConfig::get('podcast') && Access::check('interface', 75)) {
@@ -2100,7 +2063,6 @@ class Subsonic_Api
      */
     public static function downloadpodcastepisode($input)
     {
-        self::check_version($input, "1.9.0");
         $id = self::check_parameter($input, 'id');
 
         if (AmpConfig::get('podcast') && Access::check('interface', 75)) {
@@ -2125,8 +2087,6 @@ class Subsonic_Api
      */
     public static function getbookmarks($input)
     {
-        self::check_version($input, "1.9.0");
-
         $response         = Subsonic_XML_Data::createSuccessResponse();
         $bookmarks        = Bookmark::get_bookmarks();
         Subsonic_XML_Data::addBookmarks($response, $bookmarks);
@@ -2141,7 +2101,6 @@ class Subsonic_Api
      */
     public static function createbookmark($input)
     {
-        self::check_version($input, "1.9.0");
         $id       = self::check_parameter($input, 'id');
         $position = self::check_parameter($input, 'position');
         $comment  = $input['comment'];
@@ -2174,7 +2133,6 @@ class Subsonic_Api
      */
     public static function deletebookmark($input)
     {
-        self::check_version($input, "1.9.0");
         $id   = self::check_parameter($input, 'id');
         $type = Subsonic_XML_Data::getAmpacheType($id);
 
@@ -2197,8 +2155,6 @@ class Subsonic_Api
      */
     public static function getchatmessages($input)
     {
-        self::check_version($input, "1.2.0");
-
         $response = Subsonic_XML_Data::createError(Subsonic_XML_Data::SSERROR_DATA_NOTFOUND);
         self::apiOutput($input, $response);
     }
@@ -2211,8 +2167,6 @@ class Subsonic_Api
      */
     public static function addchatmessages($input)
     {
-        self::check_version($input, "1.2.0");
-
         $response = Subsonic_XML_Data::createError(Subsonic_XML_Data::SSERROR_DATA_NOTFOUND);
         self::apiOutput($input, $response);
     }
@@ -2225,8 +2179,6 @@ class Subsonic_Api
      */
     public static function getplayqueue($input)
     {
-        self::check_version($input, "1.12.0");
-
         $response = Subsonic_XML_Data::createError(Subsonic_XML_Data::SSERROR_DATA_NOTFOUND);
         self::apiOutput($input, $response);
     }
@@ -2239,8 +2191,6 @@ class Subsonic_Api
      */
     public static function saveplayqueue($input)
     {
-        self::check_version($input, "1.12.0");
-
         $response = Subsonic_XML_Data::createError(Subsonic_XML_Data::SSERROR_DATA_NOTFOUND);
         self::apiOutput($input, $response);
     }