+++ lib/class/wanted.class.php
@@ -89,16 +89,12 @@ class Wanted extends database_object
 
     /**
      * Constructor
-     * @param integer $id
+     * @param integer $wanted_id
      */
-    public function __construct($id = 0)
+    public function __construct($wanted_id)
     {
-        if (!$id) {
-            return true;
-        }
-
         /* Get the information from the db */
-        $info = $this->get_info($id);
+        $info = $this->get_info($wanted_id);
 
         // Foreach what we've got
         foreach ($info as $key => $value) {