+++ lib/class/xml_data.class.php
@@ -421,7 +421,7 @@ class XML_Data
 
             // Handle includes
             if (in_array("albums", $include)) {
-                $albums = self::albums($artist->get_albums(null, true), $include, false);
+                $albums = self::albums($artist->get_albums(), $include, false);
             } else {
                 $albums = ($artist->albums ?: 0);
             }
@@ -503,7 +503,18 @@ class XML_Data
             if (in_array("songs", $include)) {
                 $songs = self::songs($album->get_songs(), array(), false);
             } else {
-                $songs = $album->song_count;
+                if (AmpConfig::get('album_group')) {
+                    $song_count = 0;
+                    $disc_ids   = $album->get_group_disks_ids();
+                    foreach ($disc_ids as $discid) {
+                        $disc = new Album($discid);
+                        $disc->format();
+                        $song_count = $song_count + $disc->song_count;
+                    }
+                    $songs = $song_count;
+                } else {
+                    $songs = $album->song_count;
+                }
             }
 
             $string .= "\t<year>" . $album->year . "</year>\n" .