+++ docs/CHANGELOG.md
@@ -55,6 +55,11 @@ CHANGELOG
 - Add -u to bin/catalog_update.inc This function will update the artist table with bio, image, etc as well as update similar artists.
 - Numerous catalog updates to allow data migration when updating file tags. (UserActivity::migrate, Userflag::migrate, Rating::migrate, Catalog::migrate)
 - Faster tag updates/catalog verify! (Updating an album would update each file multiple times)
+- Subsonic - Update api to 1.13.0 [http://www.subsonic.org/pages/api.jsp]
+- Subsonic - Allow token auth using API Key instead of password.
+- Subsonic - Don't ignore group settings with id3 browsing
+- Subsonic - New Method: updateUser
+- Subsonic - New Method: getTopSongs
 
 3.9.1
 ----------
@@ -81,8 +86,6 @@ CHANGELOG
 - Fixed uploading of avatar.
 - DSub now writes the correct extension when the ampache config switch `encode_player_api_target` is enabled.
 - Artist info now properly displayed.
-- Fixed login bug with PHP-7.2 installed.
-
 3.8.8
 ----------
 - Subsonic API now removes illegal characters before returning text.